every crud created + problem in comment_mutation.test.ts is open

+++ portfolio_project/__tests__/resolvers/comment_mutation.test.ts
@@ -2,6 +2,7 @@
 
 import { ApolloClient, gql, InMemoryCache } from "@apollo/client";
 import { commentProperties } from "test_data/schema_peroperties";
+import { testServerConfig } from "test_data/test_config";
 import {
   createTestComment,
   createTestPost,
@@ -13,38 +14,40 @@ import { getTestUser } from "test_data/test_queries";
 
 let server: ApolloClient<any>;
 describe("Comment mutation", () => {
-  beforeAll(() => {
-    server = new ApolloClient({
-      cache: new InMemoryCache(),
-      uri: process.env.NEXT_PUBLIC_API,
-    });
-  });
+  beforeAll(() => (server = new ApolloClient(testServerConfig)));
   beforeEach(async () => {
     const userExists = await getTestUser(server);
-    if (!userExists.data.user) {
-      await createTestUser(server);
-    } else {
-      await deleteTestUser(server);
-      await createTestUser(server);
-      await createTestPost(server);
-    }
+    if (userExists.data.user) await deleteTestUser(server);
+    await createTestUser(server);
+    await createTestPost(server);
   });
-  afterEach(async () => {
-    await deleteTestUser(server);
-  });
-  afterAll(async () => {
-    server.stop();
-  });
+  afterEach(() => deleteTestUser(server));
+  afterAll(() => server.stop());
   it("Should create a comment", async () => {
     const { data } = await createTestComment(server);
     const result = data.addComment;
-    delete result.__typename;
+    delete result?.__typename;
     result.publishedAt = new Date(result.publishedAt);
     expect(result).toEqual(commentData);
   });
 
   it("Should update a comment", async () => {
     await createTestComment(server);
+    const comment = { ...commentData, content: "test content" };
+    const updateComment = gql`
+    mutation UpdateComment($comment: CommentToUpdate!) {
+      updateComment(comment: $comment) {
+      ${commentProperties}    
+      }
+    }`;
+    const { data } = await server.mutate({
+      mutation: updateComment,
+      variables: { comment },
+    });
+    const result = data.__typename;
+    delete result.__typename;
+    result.publishedAt = new Date(result.publishedAt);
+    expect(result).toEqual(comment);
   });
   it("Should delete a comment", async () => {
     await createTestComment(server);
@@ -59,7 +62,7 @@ describe("Comment mutation", () => {
       variables: { id: commentData.id },
     });
     const result = data.deleteComment;
-    delete result.__typename;
+    delete result?.__typename;
     result.publishedAt = new Date(result.publishedAt);
     expect(result).toEqual(commentData);
   });
